+++ xknx/core/__init__.py
@@ -1,6 +1,9 @@
 """Module for the automations and business logic of XKNX."""
 # flake8: noqa
+from .connection_manager import ConnectionManager
+from .connection_state import XknxConnectionState
 from .payload_reader import PayloadReader
 from .state_updater import StateUpdater
+from .task_registry import Task, TaskRegistry
 from .telegram_queue import TelegramQueue
 from .value_reader import ValueReader