+++ xknx/xknx.py
@@ -10,7 +10,13 @@ from sys import platform
 from types import TracebackType
 from typing import Awaitable, Callable
 
-from xknx.core import StateUpdater, TelegramQueue
+from xknx.core import (
+    ConnectionManager,
+    StateUpdater,
+    TaskRegistry,
+    TelegramQueue,
+    XknxConnectionState,
+)
 from xknx.devices import Device, Devices
 from xknx.io import (
     DEFAULT_MCAST_GRP,
@@ -37,6 +43,8 @@ class XKNX:
         address_format: GroupAddressType = GroupAddressType.LONG,
         telegram_received_cb: Callable[[Telegram], Awaitable[None]] | None = None,
         device_updated_cb: Callable[[Device], Awaitable[None]] | None = None,
+        connection_state_changed_cb: Callable[[XknxConnectionState], Awaitable[None]]
+        | None = None,
         rate_limit: int = DEFAULT_RATE_LIMIT,
         multicast_group: str = DEFAULT_MCAST_GRP,
         multicast_port: int = DEFAULT_MCAST_PORT,
@@ -51,16 +59,17 @@ class XKNX:
         self.sigint_received = asyncio.Event()
         self.telegram_queue = TelegramQueue(self)
         self.state_updater = StateUpdater(self)
+        self.connection_manager = ConnectionManager()
+        self.task_registry = TaskRegistry(self)
+        self.start_state_updater = state_updater
         self.knxip_interface: KNXIPInterface | None = None
         self.started = asyncio.Event()
-        self.connected = asyncio.Event()
         self.address_format = address_format
         self.own_address = IndividualAddress(own_address)
         self.rate_limit = rate_limit
         self.multicast_group = multicast_group
         self.multicast_port = multicast_port
         self.connection_config = connection_config
-        self.start_state_updater = state_updater
         self.daemon_mode = daemon_mode
         self.version = VERSION
 
@@ -72,6 +81,11 @@ class XKNX:
 
         if device_updated_cb is not None:
             self.devices.register_device_updated_cb(device_updated_cb)
+
+        if connection_state_changed_cb is not None:
+            self.connection_manager.register_connection_state_changed_cb(
+                connection_state_changed_cb
+            )
 
     def __del__(self) -> None:
         """Destructor. Cleaning up if this was not done before."""
@@ -98,6 +112,7 @@ class XKNX:
 
     async def start(self) -> None:
         """Start XKNX module. Connect to KNX/IP devices and start state updater."""
+        self.task_registry.start()
         self.knxip_interface = KNXIPInterface(
             self, connection_config=self.connection_config
         )
@@ -126,6 +141,7 @@ class XKNX:
 
     async def stop(self) -> None:
         """Stop XKNX module."""
+        self.task_registry.stop()
         self.state_updater.stop()
         await self.join()
         await self.telegram_queue.stop()